requirements.txt
hikari
python-dotenv
